+++ src/discord/membercommands.py
@@ -16,7 +16,7 @@ from src.wiki.scilympiad import make_results_template, get_points
 from src.wiki.schools import get_school_listing
 from src.mongo.mongo import insert
 from commands import get_list, get_quick_list, get_help
-from lists import get_state_list
+from src.lists import get_state_list
 from src.discord.utils import lookup_role
 from src.discord.mute import _mute
 from commandchecks import is_staff